+++ plugins/istio/install.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+ergo ops wget https://github.com/istio/istio/releases/download/1.11.6/istio-1.11.6-linux-amd64.tar.gz
+tar xzvfC ~/.ergo/cache/istio-1.11.6-linux-amd64.tar.gz /tmp
+mv /tmp/istio-1.11.6/bin/istioctl /usr/local/bin
+rm -rf ~/.ergo/cache/istio-1.11.6-linux-amd64.tar.gz