+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 > 记录版本变更
 
+## 2022021017
+
+- 新增`helm`, `istio`, `cilium`
+
 ## 2022020815
 
 - 新增consul、mariadb、redis、mongodb、mysql、postgresql